Weapon Changes
1.) Created a weapons scriptableObject and a WeaponController script. Also following a template for how all the weapon prefabs are structured to aid modularity.

2.) Added more variables such as bloom, and numOfProjectiles to allow a shotgun weapon.

3.) Added sound effects for reloading and dry triggers

4.) Added the concept of ammo and clipSize so guns don't all have unlimited ammo.

+++ Man Vs Metal Prototype/Assets/Damage/Damageable.cs
@@ -7,14 +7,14 @@ public class Damageable : MonoBehaviour
 {
     [SerializeField] private int startingHealth;
 
-    private int health;
+    private float health;
 
     private void Start()
     {
         health = startingHealth;
     }
 
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         health -= damageAmount;
         if (health < 0)
